Coordinates refer to locations in memory. Changed to size_t. Proposed knockBack function signature?

+++ src/Map.h
@@ -42,8 +42,8 @@ struct Coord
  		return temp;
  	}
 	
-    unsigned int y;
-    unsigned int x;
+    size_t y;
+    size_t x;
 };
 
 class Piece;
